move mime-types request into holder
