+++ aries_cloudagent/holder/routes.py
@@ -19,6 +19,10 @@ from ..messaging.valid import (
 from ..wallet.error import WalletNotFoundError
 
 
+class AttributeMimeTypesResultSchema(Schema):
+    """Result schema for credential attribute MIME type."""
+
+
 class RawEncCredAttrSchema(Schema):
     """Credential attribute schema."""
 
@@ -120,6 +124,27 @@ async def credentials_get(request: web.BaseRequest):
     return web.json_response(credential_json)
 
 
+@docs(tags=["credentials"], summary="Get attribute MIME types from wallet")
+@match_info_schema(CredIdMatchInfoSchema())
+@response_schema(AttributeMimeTypesResultSchema(), 200)
+async def credentials_attr_mime_types_get(request: web.BaseRequest):
+    """
+    Request handler for getting credential attribute MIME types.
+
+    Args:
+        request: aiohttp request object
+
+    Returns:
+        The MIME types response
+
+    """
+    context = request.app["request_context"]
+    credential_id = request.match_info["credential_id"]
+    holder: BaseHolder = await context.inject(BaseHolder)
+
+    return web.json_response(await holder.get_mime_type(credential_id))
+
+
 @docs(tags=["credentials"], summary="Remove a credential from the wallet by id")
 @match_info_schema(CredIdMatchInfoSchema())
 async def credentials_remove(request: web.BaseRequest):
@@ -190,6 +215,11 @@ async def register(app: web.Application):
     app.add_routes(
         [
             web.get("/credential/{credential_id}", credentials_get, allow_head=False),
+            web.get(
+                "/issue-credential/mime-types/{credential_id}",
+                credentials_attr_mime_types_get,
+                allow_head=False,
+            ),
             web.post("/credential/{credential_id}/remove", credentials_remove),
             web.get("/credentials", credentials_list, allow_head=False),
         ]
